[#12218] refactor(secret): Extract SecretWriteContext interface
Keep the SPI on an interface and move the previous concrete type to
ServiceSecretWriteContext for server-side write-through.
